Сущности user и post были вынесены в отдельные файлы
