+++ post.py
@@ -1,0 +1,5 @@
+class Post:
+    def __init__(self, post_id=0, text="", photo=""):
+        self.post_id = post_id
+        self.text = text
+        self.photo = photo
