+++ user.py
@@ -1,0 +1,5 @@
+class User:
+    def __init__(self, name, tg_id, admin=False):
+        self.name = name
+        self.tg_id = tg_id
+        self.admin = admin
