+++ vkparser.py
@@ -1,12 +1,6 @@
 import requests
 from config import Config
-
-
-class Post:
-    def __init__(self, post_id=0, text="", photo=""):
-        self.post_id = post_id
-        self.text = text
-        self.photo = photo
+from post import Post
 
 
 class VkWallParser:
